requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary>=2.9.9
rapidfuzz==3.5.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3

